Make example override existing input
